requirements.txt
pandas>=2.2
numpy>=1.26
Pillow>=10.3
timm>=0.9.16
torch>=2.0.0
torchvision>=0.15.1
pytorch-grad-cam>=1.4.8
pyyaml>=6.0
tqdm>=4.66
opencv-python-headless>=4.10
